+++ regen/memory/aligned-array.h
