+++ regen/shapes/bounding-sphere.h
@@ -4,7 +4,7 @@
 #include <regen/shapes/bounding-shape.h>
 #include <regen/shapes/aabb.h>
 #include <regen/shapes/obb.h>
-#include <regen/utility/aligned-array.h>
+#include <regen/memory/aligned-array.h>
 #include "batch-of-shapes.h"
 
 namespace regen {
