+++ regen/shapes/obb.h
@@ -3,7 +3,7 @@
 
 #include <regen/shapes/bounding-box.h>
 #include <regen/shapes/aabb.h>
-#include <regen/utility/aligned-array.h>
+#include <regen/memory/aligned-array.h>
 #include "batch-of-shapes.h"
 
 namespace regen {
